pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.121.2",
    "google-genai>=1.50.1",
    "google-generativeai>=0.8.5",
    "passlib>=1.7.4",
    "pillow>=12.0.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]
